+++ tap_jira/__init__.py
@@ -31,31 +31,39 @@ def test_credentials_are_authorized(config):
     client.request(streams_.ISSUES.tap_stream_id, "GET", "/rest/api/2/search",
                    params={"maxResults": 1})
 
-
-def load_metadata(stream, schema):
-    mdata = metadata.new()
-
-    for field_name in schema.properties.keys():
-        if field_name in stream.pk_fields:
-            mdata = metadata.write(mdata, ('properties', field_name), 'inclusion', 'automatic')
-        else:
-            mdata = metadata.write(mdata, ('properties', field_name), 'inclusion', 'available')
-
-    return metadata.to_list(mdata)
-
 def discover(config):
     test_credentials_are_authorized(config)
     catalog = Catalog([])
     for stream in streams_.all_streams:
         schema = Schema.from_dict(load_schema(stream.tap_stream_id))
+
+        mdata = generate_metadata(stream, schema)
+        
         catalog.streams.append(CatalogEntry(
             stream=stream.tap_stream_id,
             tap_stream_id=stream.tap_stream_id,
             key_properties=stream.pk_fields,
             schema=schema,
-            metadata=load_metadata(stream, schema)
-        ))
+
+            metadata=mdata))
     return catalog
+
+
+def generate_metadata(stream, schema):
+    mdata = metadata.new()
+    mdata = metadata.write(mdata, (), 'table-key-properties', stream.pk_fields)
+    #mdata = metadata.write(mdata, (), 'forced-replication-method', stream.replication_method)
+
+    #if stream.replication_key:
+    #    mdata = metadata.write(mdata, (), 'valid-replication-keys', [stream.replication_key])
+
+    for field_name in schema.properties.keys():
+        if field_name in stream.pk_fields: #or field_name == stream.replication_key:
+            mdata = metadata.write(mdata, ('properties', field_name), 'inclusion', 'automatic')
+        else:
+            mdata = metadata.write(mdata, ('properties', field_name), 'inclusion', 'available')
+
+    return metadata.to_list(mdata)
 
 
 def output_schema(stream):
@@ -63,29 +71,32 @@ def output_schema(stream):
     singer.write_schema(stream.tap_stream_id, schema, stream.pk_fields)
 
 
-def sync(ctx):
-    streams_.validate_dependencies(ctx)
-    currently_syncing = ctx.state.get("currently_syncing")
+def sync():
+    streams_.validate_dependencies()
+    currently_syncing = Context.state.get("currently_syncing")
     start_idx = streams_.all_stream_ids.index(currently_syncing) \
         if currently_syncing else 0
-    streams = [s for s in streams_.all_streams[start_idx:]
-               if s.tap_stream_id in ctx.selected_stream_ids]
+
     # two loops through streams are necessary so that the schema is output
     # BEFORE syncing any streams. Otherwise, the first stream might generate
     # data for the second stream, but the second stream hasn't output its
     # schema yet
-    for stream in streams:
+    for stream in streams_.all_streams:
         output_schema(stream)
-    for stream in streams:
+
+    for stream in streams_.all_streams:
+        if not Context.is_selected(stream.tap_stream_id):
+            continue
+        
         # indirect_stream indicates the data for the stream comes from some
         # other stream, so we don't sync it directly.
         if stream.indirect_stream:
             continue
-        ctx.state["currently_syncing"] = stream.tap_stream_id
-        ctx.write_state()
-        stream.sync(ctx)
-    ctx.state["currently_syncing"] = None
-    ctx.write_state()
+        Context.state["currently_syncing"] = stream.tap_stream_id
+        Context.write_state()
+        stream.sync()
+    Context.state["currently_syncing"] = None
+    Context.write_state()
 
 
 def main_impl():
@@ -96,7 +107,11 @@ def main_impl():
     else:
         catalog = Catalog.from_dict(args.properties) \
             if args.properties else discover(args.config)
-        sync(Context(args.config, args.state, catalog))
+        Context.config = args.config
+        Context.state = args.state
+        Context.catalog = catalog
+        Context.client = Client(Context.config)
+        sync()
 
 
 def main():