+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
 
+## 0.3.3
+  * Update version of `requests` to `2.20.0` in response to CVE 2018-18074
+
+## 0.3.2
+  * Changes `search` endpoint for issues to use the user's timezone [#15](https://github.com/singer-io/tap-jira/pull/15)
+
+## 0.2.3
+  * Added `jsdPublic` field to `ticket_comments` schema.
+
 ## 0.2.2
   * Fixes a pagination bug that would cause Issue and Versions to return more pages than necessary [#14](https://github.com/singer-io/tap-jira/pull/14)